move legalentities, transferinstruments and document endpoints to the new legalEntityManagement API
