+++ lib/adyen-ruby-api-library.rb
@@ -15,3 +15,4 @@ require_relative "adyen/services/balance_platform"
 require_relative "adyen/hash_with_accessors"
 require_relative "adyen/utils/hmac_validator"
 require_relative "adyen/services/transfers"
+require_relative "adyen/services/legal_entity_management"
