+++ spec/mocks/requests/LegalEntityManagement/create_legal_entity.json
