+++ spec/mocks/requests/LegalEntityManagement/create_transfer_instrument.json
