+++ spec/mocks/requests/LegalEntityManagement/update_legal_entity.json
