+++ spec/mocks/responses/LegalEntityManagement/create_document.json
