+++ spec/mocks/responses/LegalEntityManagement/create_legal_entity.json
