+++ spec/mocks/responses/LegalEntityManagement/create_transfer_instrument.json
