+++ spec/mocks/responses/LegalEntityManagement/get_transfer_instrument.json
