+++ spec/mocks/responses/LegalEntityManagement/update_document.json
@@ -12,6 +12,7 @@
       }
   ],
   "description": "passportimagess",
+  "fileName": "passportimagess",
   "owner": {
       "id": "LE322JV223222D5DTCP6R2Q57",
       "type": "legalEntity"
