+++ spec/mocks/responses/LegalEntityManagement/update_legal_entity.json
@@ -16,6 +16,7 @@
         "street":"Another Street",
         "street2":"Another Building"
         },
+        "taxExempt":"false",
         "taxId":"101002749",
         "type":"privateCompany"},
   "type": "organization"
