+++ spec/mocks/responses/LegalEntityManagement/update_transfer_instrument.json
